chore: refactor rari crate by removing rari_rsc module and integrating its functionality directly into the main crate
