+++ crates/rari/Cargo.toml
@@ -87,7 +87,6 @@ async-compression = { version = "0.4.42", features = [
 ] }
 
 # === Serialization ===
-rari_rsc = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 base64 = { workspace = true }
