--- crates/rari_rsc/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "rari_rsc"
-edition = { workspace = true }
-authors = { workspace = true }
-license = { workspace = true }
-repository = { workspace = true }
-publish = false
-
-[lib]
-crate-type = [ "rlib" ]
-
-[dependencies]
-cow-utils = { workspace = true }
-parking_lot = { workspace = true }
-regex = { workspace = true }
-rustc-hash = { workspace = true }
-serde = { workspace = true }
-smallvec = { workspace = true }
-
-[lints]
-workspace = true
