--- crates/rari_rsc/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod components;
-pub mod utils;
-
-pub use components::{
-    ComponentContext, ComponentProp, ComponentRegistry, ComponentType, TransformedComponent,
-};
-pub use utils::extract_dependencies;
